\name{plotdiag}
\alias{plotdiag}

\title{ Diagnostic Plots for SDDA }

\description{Diagnostic Plots for SDDA to check the choice of the number
  of variables.}

\usage{plotdiag(obj)}

\arguments{
  \item{obj}{Output from \code{sdda} }
}

\author{ Glenn Stone }
\keyword{models}
\keyword{multivariate}
