\name{print.sdda}
\alias{print.sdda}

\title{ Print SDDA output to screen }

\description{ Summary statistics for an SDDA object are printed to the
  screen including the number of samples, the number of variables, the
  number of groups and their names etc. Pass plot=TRUE to get plot
  diagnostics also.  }

\usage{
print.sdda(x,plot=FALSE,...)
}

\arguments{
  \item{x}{output from SDDA}
  \item{plot}{Should we plot some diagnostics?}
  \item{\dots}{additional parameters as needed }
}

\author{ David Clifford }

\keyword{models}% at least one, from doc/KEYWORDS
\keyword{multivariate}% __ONLY ONE__ keyword per line
