\name{summary.sdda}
\alias{summary.sdda}

\title{ Summarise SDDA output }

\description{Summarise output from SDDA }

\usage{ summary.sdda(object, ...)  }

\arguments{
  \item{object}{ Output from \code{sdda} }
  \item{\dots}{ Additional parameters as needed }
}

\author{ David Clifford }

\keyword{models}
\keyword{multivariate}
