\name{which.genes}
\alias{which.genes}

\title{Print list of genes chosen by SDDA}
\description{This function processes the sdda output to list the genes chosen.}

\usage{which.genes(obj) }

\arguments{
  \item{obj}{ output from \code{sdda} }
}

\author{Glenn Stone}

\keyword{models}
\keyword{multivariate}
